refactor: transfer & multipayment in single modal (#1904)
