+++ src/renderer/components/Transaction/TransactionForm/TransactionFormTransfer.vue
@@ -26,7 +26,7 @@
     <WalletSelection
       v-if="schema && schema.address"
       v-model="$v.wallet.$model"
-      :compatible-address="$v.form.recipientId.$model"
+      :compatible-address="$v.recipientId.$model"
       class="TransactionFormTransfer__wallet mb-5"
       profile-class="mb-5"
       @select="ensureAvailableAmount"
@@ -34,11 +34,12 @@
 
     <InputAddress
       ref="recipient"
-      v-model="$v.form.recipientId.$model"
+      v-model="$v.recipientId.$model"
       :label="$t('TRANSACTION.RECIPIENT')"
       :pub-key-hash="walletNetwork.version"
       :show-suggestions="true"
       :is-disabled="!currentWallet"
+      :warning-text="recipientWarning"
       name="recipientId"
       class="TransactionFormTransfer__recipient mb-5"
     />
@@ -46,10 +47,10 @@
     <div class="flex items-baseline mb-5">
       <InputCurrency
         ref="amount"
-        v-model="$v.form.amount.$model"
+        v-model="$v.amount.$model"
         :alternative-currency="alternativeCurrency"
         :currency="walletNetwork.token"
-        :is-invalid="$v.form.amount.$dirty && $v.form.amount.$invalid"
+        :is-invalid="$v.amount.$dirty && $v.amount.$invalid"
         :label="$t('TRANSACTION.AMOUNT')"
         :minimum-error="amountTooLowError"
         :minimum-amount="minimumAmount"
@@ -63,14 +64,33 @@
         @input="setSendAll(false, false)"
       />
 
+      <ButtonGeneric
+        v-if="hasAip11"
+        :disabled="!isValidRecipient"
+        :label="$t('TRANSACTION.BUTTON_ADD')"
+        class="TransactionFormTransfer__add py-1 flex-inline h-8 mt-4 mr-3"
+        @click="addRecipient"
+      />
+
       <InputSwitch
         v-model="isSendAllActive"
         :text="$t('TRANSACTION.SEND_ALL')"
-        :is-disabled="!canSendAll || !currentWallet"
+        :is-disabled="isRecipientSendAll || !canSendAll || !currentWallet"
         class="TransactionFormTransfer__send-all"
         @change="setSendAll"
       />
     </div>
+
+    <TransactionRecipientList
+      v-if="hasAip11"
+      :items="$v.form.recipients.$model"
+      :max-items="maximumRecipients"
+      :show-count="true"
+      :is-invalid="hasMoreThanMaximumRecipients"
+      :required="true"
+      class="TransactionFormTransfer__recipients mt-4"
+      @remove="emitRemoveRecipient"
+    />
 
     <InputText
       ref="vendorField"
@@ -78,26 +98,36 @@
       :label="vendorFieldLabel"
       :bip39-warning="true"
       :helper-text="vendorFieldHelperText"
-      :is-disabled="!currentWallet"
       :maxlength="vendorFieldMaxLength"
       name="vendorField"
       class="TransactionFormTransfer__vendorfield mb-5"
     />
 
-    <InputFee
-      ref="fee"
-      :currency="walletNetwork.token"
-      :transaction-type="$options.transactionType"
-      :is-disabled="!currentWallet"
-      :wallet="currentWallet"
-      :wallet-network="walletNetwork"
-      class="TransactionFormTransfer__fee"
-      @input="onFee"
-    />
+    <div class="mt-4">
+      <InputFee
+        ref="fee"
+        :currency="walletNetwork.token"
+        :transaction-type="transactionTypeFee"
+        :is-disabled="!currentWallet"
+        :wallet="currentWallet"
+        :wallet-network="walletNetwork"
+        class="TransactionFormTransfer__fee"
+        :class="{
+          'TransactionFormTransfer__fee--helper': isMultiPayment && insufficientFundsError
+        }"
+        @input="onFee"
+      />
+      <p
+        v-if="isMultiPayment && insufficientFundsError"
+        class="text-red-dark text-theme-page-text-light text-xs"
+      >
+        {{ insufficientFundsError }}
+      </p>
+    </div>
 
     <div v-if="!isMultiSignature">
       <div
-        v-if="currentWallet && currentWallet.isLedger"
+        v-if="isLedger"
         class="TransactionFormTransfer__ledger-notice mt-10"
       >
         {{ $t('TRANSACTION.LEDGER_SIGN_NOTICE') }}
@@ -136,9 +166,9 @@
     <footer class="mt-10 flex justify-between items-center">
       <div class="self-start">
         <button
-          :disabled="$v.form.$invalid"
+          :disabled="!isFormValid"
           class="TransactionFormTransfer__next blue-button"
-          @click="onSubmit"
+          @click="nextStep"
         >
           {{ $t('COMMON.NEXT') }}
         </button>
@@ -185,11 +215,13 @@
 <script>
 import { required } from 'vuelidate/lib/validators'
 import { TRANSACTION_TYPES, VENDOR_FIELD } from '@config'
+import { ButtonGeneric } from '@/components/Button'
 import { InputAddress, InputCurrency, InputPassword, InputSwitch, InputText, InputFee } from '@/components/Input'
 import { ListDivided, ListDividedItem } from '@/components/ListDivided'
 import { ModalConfirmation, ModalLoader } from '@/components/Modal'
 import { PassphraseInput } from '@/components/Passphrase'
 import SvgIcon from '@/components/SvgIcon'
+import TransactionRecipientList from '@/components/Transaction/TransactionRecipientList'
 import WalletSelection from '@/components/Wallet/WalletSelection'
 import WalletService from '@/services/wallet'
 import mixin from './mixin'
@@ -200,6 +232,7 @@ export default {
   transactionType: TRANSACTION_TYPES.GROUP_1.TRANSFER,
 
   components: {
+    ButtonGeneric,
     InputAddress,
     InputCurrency,
     InputPassword,
@@ -212,6 +245,7 @@ export default {
     ModalLoader,
     PassphraseInput,
     SvgIcon,
+    TransactionRecipientList,
     WalletSelection
   },
 
@@ -226,23 +260,64 @@ export default {
   },
 
   data: () => ({
+    recipientId: '',
+    amount: '',
+    isSendAllActive: false,
+    previousAmount: '',
+    showConfirmSendAll: false,
+    wallet: null,
     form: {
-      amount: '',
+      recipients: [],
       fee: 0,
       passphrase: '',
       walletPassword: '',
-      recipientId: '',
       vendorField: ''
-    },
-    isSendAllActive: false,
-    previousAmount: '',
-    wallet: null,
-    showConfirmSendAll: false
+    }
   }),
 
   computed: {
     alternativeCurrency () {
       return this.$store.getters['session/currency']
+    },
+
+    isValidRecipient () {
+      if (!this.$v.recipientId.$dirty || this.$v.recipientId.$invalid) {
+        return false
+      }
+
+      if (!this.$v.amount.$dirty || this.$v.amount.$invalid) {
+        return false
+      }
+
+      return true
+    },
+
+    isFormValid () {
+      if (this.isMultiPayment) {
+        return !this.$v.form.$invalid && !this.insufficientFundsError
+      }
+
+      return !this.$v.form.$invalid
+    },
+
+    isMultiPayment () {
+      return this.form.recipients.length > 1
+    },
+
+    isLedger () {
+      return this.currentWallet && !!this.currentWallet.isLedger
+    },
+
+    hasAip11 () {
+      return this.walletNetwork.constants && !!this.walletNetwork.constants.aip11 && !this.isLedger
+    },
+
+    hasMoreThanMaximumRecipients () {
+      return this.form.recipients.length > this.maximumRecipients
+    },
+
+    transactionTypeFee () {
+      return this.isMultiPayment ? TRANSACTION_TYPES.GROUP_1.MULTI_PAYMENT : TRANSACTION_TYPES.GROUP_1.TRANSFER
     },
 
     // Customize the message to display the minimum amount as subunit
@@ -274,7 +349,49 @@ export default {
         return this.currency_subToUnit(0)
       }
 
-      return this.currency_subToUnit(this.currentWallet.balance).minus(this.form.fee)
+      if (!this.hasAip11) {
+        return this.currency_subToUnit(this.currentWallet.balance).minus(this.form.fee)
+      }
+
+      let availableAmount = this.currency_subToUnit(this.currentWallet.balance).minus(this.form.fee)
+
+      for (const recipient of this.form.recipients) {
+        if (!recipient.sendAll) {
+          availableAmount = availableAmount.minus(this.currency_subToUnit(recipient.amount))
+        }
+      }
+
+      return availableAmount
+    },
+
+    totalAmount () {
+      const amount = this.currency_toBuilder(0)
+
+      for (const recipient of this.$v.form.recipients.$model) {
+        amount.add(recipient.amount)
+      }
+
+      return amount.value
+    },
+
+    insufficientFundsError () {
+      if (!this.currentWallet) {
+        return null
+      }
+
+      const funds = this.currency_unitToSub(this.currency_subToUnit(this.currentWallet.balance))
+      const totalAmount = this.currency_unitToSub(this.currency_subToUnit(this.totalAmount).plus(this.form.fee))
+
+      if (funds.isLessThan(totalAmount)) {
+        const balance = this.formatter_networkCurrency(this.currentWallet.balance)
+        return this.$t('TRANSACTION.ERROR.NOT_ENOUGH_BALANCE', { balance })
+      }
+
+      return null
+    },
+
+    isRecipientSendAll () {
+      return this.form.recipients.some(recipient => !!recipient.sendAll)
     },
 
     canSendAll () {
@@ -340,13 +457,34 @@ export default {
       }
 
       return VENDOR_FIELD.defaultMaxLength
+    },
+
+    recipientWarning () {
+      if (!this.$v.recipientId.$dirty) {
+        return null
+      }
+
+      if (this.form.recipients.some(recipient => recipient.address === this.$v.recipientId.$model)) {
+        return this.$t('TRANSACTION.MULTI_PAYMENT.WARNING_DUPLICATE')
+      }
+
+      return null
+    },
+
+    maximumRecipients () {
+      if (!this.session_network.constants || !this.session_network.constants.multiPaymentLimit) {
+        return 500
+      }
+
+      return this.session_network.constants.multiPaymentLimit
     }
   },
 
   watch: {
     wallet () {
       this.ensureAvailableAmount()
-      this.$v.form.recipientId.$touch()
+      this.$v.recipientId.$touch()
+      this.$v.amount.$touch()
     }
   },
 
@@ -363,15 +501,26 @@ export default {
     getTransactionData () {
       const transactionData = {
         address: this.currentWallet.address,
-        amount: this.currency_unitToSub(this.form.amount),
-        recipientId: this.form.recipientId,
         vendorField: this.form.vendorField,
         passphrase: this.form.passphrase,
         fee: this.getFee(),
         wif: this.form.wif,
         networkWif: this.walletNetwork.wif,
-        networkId: this.walletNetwork.id,
         multiSignature: this.currentWallet.multiSignature
+      }
+
+      if (this.isMultiPayment) {
+        transactionData.recipients = this.form.recipients
+      } else {
+        if (this.hasAip11) {
+          transactionData.recipientId = this.form.recipients[0].address
+          transactionData.amount = this.getAmountNormalTransaction()
+        } else {
+          transactionData.recipientId = this.recipientId
+          transactionData.amount = this.currency_unitToSub(this.amount)
+        }
+
+        transactionData.networkId = this.walletNetwork.id
       }
 
       if (this.currentWallet.secondPublicKey) {
@@ -381,8 +530,28 @@ export default {
       return transactionData
     },
 
+    getAmountNormalTransaction () {
+      if (this.insufficientFundsError) {
+        return this.currency_unitToSub(this.currency_subToUnit(this.form.recipients[0].amount).minus(this.form.fee))
+      }
+
+      return this.form.recipients[0].amount
+    },
+
     async buildTransaction (transactionData, isAdvancedFee = false, returnObject = false) {
-      return this.$client.buildTransfer(transactionData, isAdvancedFee, returnObject)
+      if (this.isMultiPayment) {
+        return this.$client.buildMultiPayment(transactionData, isAdvancedFee, returnObject)
+      } else {
+        return this.$client.buildTransfer(transactionData, isAdvancedFee, returnObject)
+      }
+    },
+
+    transactionError () {
+      if (this.isMultiPayment) {
+        this.$error(this.$t('TRANSACTION.ERROR.VALIDATION.MULTI_PAYMENT'))
+      } else {
+        this.$error(this.$t('TRANSACTION.ERROR.VALIDATION.TRANSFER'))
+      }
     },
 
     populateSchema () {
@@ -390,9 +559,10 @@ export default {
         return
       }
 
-      this.$set(this.form, 'amount', this.schema.amount || '')
-      this.$set(this.form, 'recipientId', this.schema.address || '')
+      this.$set(this, 'amount', this.schema.amount || '')
+      this.$set(this, 'recipientId', this.schema.address || '')
       this.$set(this.form, 'vendorField', this.schema.vendorField || '')
+
       if (this.schema.wallet) {
         const currentProfileId = this.$store.getters['session/profileId']
         const ledgerWallets = this.$store.getters['ledger/isConnected'] ? this.$store.getters['ledger/wallets'] : []
@@ -442,30 +612,48 @@ export default {
       }
     },
 
-    transactionError () {
-      this.$error(this.$t('TRANSACTION.ERROR.VALIDATION.TRANSFER'))
-    },
+    addRecipient () {
+      if (this.$v.recipientId.$invalid || this.$v.amount.$invalid) {
+        return
+      }
 
-    emitNext (transaction) {
-      this.$emit('next', {
-        transaction,
-        wallet: this.senderWallet
+      this.$v.form.recipients.$model.push({
+        address: this.recipientId,
+        amount: this.isSendAllActive ? 0 : this.currency_unitToSub(this.amount),
+        sendAll: this.isSendAllActive
       })
+
+      if (this.isSendAllActive) {
+        this.previousAmount = ''
+        this.isSendAllActive = false
+      }
+
+      this.$refs.recipient.reset()
+      this.$v.recipientId.$reset()
+      this.$refs.amount.reset()
+      this.$v.amount.$model = ''
     },
 
-    onFee (fee) {
-      this.$set(this.form, 'fee', fee)
-      this.ensureAvailableAmount()
+    emitRemoveRecipient (index) {
+      if (!Object.prototype.hasOwnProperty.call(this.$v.form.recipients.$model, index)) {
+        return
+      }
+
+      this.$v.form.recipients.$model = [
+        ...this.form.recipients.slice(0, index),
+        ...this.form.recipients.slice(index + 1)
+      ]
     },
 
     setSendAll (isActive, setPreviousAmount = true) {
       if (isActive) {
         this.confirmSendAll()
-        this.previousAmount = this.form.amount
+        this.previousAmount = this.amount
       } else {
         if (setPreviousAmount && !!this.previousAmount) {
-          this.$set(this.form, 'amount', this.previousAmount)
+          this.$v.amount.$model = this.previousAmount
         }
+
         this.previousAmount = ''
         this.isSendAllActive = isActive
         this.ensureAvailableAmount()
@@ -474,7 +662,7 @@ export default {
 
     ensureAvailableAmount () {
       if (this.isSendAllActive && this.canSendAll) {
-        this.$set(this.form, 'amount', this.maximumAvailableAmount)
+        this.$v.amount.$model = this.maximumAvailableAmount
       }
     },
 
@@ -493,6 +681,31 @@ export default {
       this.isSendAllActive = false
     },
 
+    async nextStep () {
+      if (this.hasAip11) {
+        for await (const recipient of this.$v.form.recipients.$model) {
+          if (recipient.sendAll) {
+            recipient.amount = this.currency_unitToSub(this.maximumAvailableAmount)
+            break
+          }
+        }
+      }
+
+      this.onSubmit()
+    },
+
+    onFee (fee) {
+      this.$v.form.fee.$model = fee
+      this.ensureAvailableAmount()
+    },
+
+    emitNext (transaction) {
+      this.$emit('next', {
+        transaction,
+        wallet: this.senderWallet
+      })
+    },
+
     async loadTransaction () {
       try {
         const raw = await this.electron_readFile()
@@ -501,22 +714,43 @@ export default {
           try {
             const transaction = JSON.parse(raw)
 
-            if (parseInt(transaction.type, 10) !== TRANSACTION_TYPES.GROUP_1.TRANSFER) {
+            if (parseInt(transaction.type, 10) !== TRANSACTION_TYPES.GROUP_1.TRANSFER && parseInt(transaction.type, 10) !== TRANSACTION_TYPES.GROUP_1.MULTI_PAYMENT) {
               throw new Error(this.$t('VALIDATION.INVALID_TYPE'))
             }
 
-            if (transaction.recipientId) {
+            if (this.hasAip11 && Object.prototype.hasOwnProperty.call(transaction, 'asset') && Object.prototype.hasOwnProperty.call(transaction.asset, 'payments')) {
+              this.$refs.recipient.reset()
+              this.$refs.amount.reset()
+              this.$v.form.recipients.$model = []
+
+              transaction.asset.payments.forEach(payment => {
+                if (payment.recipientId && payment.amount) {
+                  if (WalletService.validateAddress(payment.recipientId, this.session_network.version)) {
+                    const amount = this.currency_unitToSub(this.currency_subToUnit(payment.amount, this.session_network), this.session_network)
+
+                    this.$v.form.recipients.$model.push({
+                      address: payment.recipientId,
+                      amount
+                    })
+                  } else {
+                    throw new Error(this.$t('VALIDATION.RECIPIENT_DIFFERENT_NETWORK', [
+                      this.wallet_truncate(payment.recipientId)
+                    ]))
+                  }
+                }
+              })
+            } else if (transaction.recipientId && transaction.amount) {
               if (WalletService.validateAddress(transaction.recipientId, this.session_network.version)) {
+                this.$v.form.recipients.$model = []
                 this.$refs.recipient.model = transaction.recipientId
+                this.$refs.amount.model = this.currency_subToUnit(transaction.amount, this.session_network)
               } else {
                 throw new Error(this.$t('VALIDATION.RECIPIENT_DIFFERENT_NETWORK', [
                   this.wallet_truncate(transaction.recipientId)
                 ]))
               }
-            }
-
-            if (transaction.amount) {
-              this.$refs.amount.model = this.currency_subToUnit(transaction.amount, this.session_network)
+            } else {
+              throw new Error(this.$t('VALIDATION.INVALID_FORMAT'))
             }
 
             if (transaction.fee) {
@@ -543,41 +777,64 @@ export default {
   },
 
   validations: {
+    recipientId: {
+      required,
+      isValid () {
+        if (this.$refs.recipient) {
+          return !this.$refs.recipient.$v.$invalid
+        }
+
+        return false
+      }
+    },
+    amount: {
+      required,
+      isValid () {
+        if (this.$refs.amount) {
+          return !this.$refs.amount.$v.$invalid
+        }
+
+        return false
+      }
+    },
     wallet: {},
     form: {
+      recipients: {
+        aboveMinimum () {
+          if (!this.hasAip11) {
+            return true
+          }
+
+          return this.form.recipients.length >= 1
+        },
+        belowOrEqualMaximum () {
+          return this.form.recipients.length <= this.maximumRecipients
+        }
+      },
       fee: mixin.validators.fee,
       passphrase: mixin.validators.passphrase,
       secondPassphrase: mixin.validators.secondPassphrase,
       vendorField: {},
-      walletPassword: mixin.validators.walletPassword,
-
-      recipientId: {
-        required,
-        isValid () {
-          if (this.$refs.recipient) {
-            return !this.$refs.recipient.$v.$invalid
-          }
-          return false
-        }
-      },
-
-      amount: {
-        required,
-        isValid () {
-          if (this.$refs.amount) {
-            return !this.$refs.amount.$v.$invalid
-          }
-          return false
-        }
-      }
+      walletPassword: mixin.validators.walletPassword
     }
   }
 }
 </script>
 
 <style>
+.TransactionFormTransfer__recipients .InputEditableList__list {
+  max-height: 13rem;
+}
+
+.TransactionFormTransfer__fee.TransactionFormTransfer__fee--helper {
+  margin-bottom: 0;
+}
+.TransactionFormTransfer__fee--helper .InputField__helper {
+  display: none;
+}
+
 .SendAllConfirmation .ModalConfirmation__container {
   min-width: calc(var(--contact-identicon-xl) + 74px * 2);
-  max-width: calc(var(--contact-identicon-xl) + 74px * 2 + 50px)
+  max-width: calc(var(--contact-identicon-xl) + 74px * 2 + 50px);
 }
 </style>
