+++ src/renderer/components/Transaction/TransactionModal.vue
@@ -387,6 +387,7 @@ export default {
 .TransactionModalTransfer {
   /* To allow more space on the fee slider */
   min-width: 38rem;
+  max-height: 100%;
 }
 
 .TransactionModalBridgechainRegistration,
@@ -394,7 +395,6 @@ export default {
 .TransactionModalBusinessRegistration,
 .TransactionModalBusinessUpdate,
 .TransactionModalIpfs,
-.TransactionModalMultiPayment,
 .TransactionModalMultiSignature {
   min-width: 35rem;
   max-height: 80vh;
