+++ src/renderer/components/Transaction/TransactionRecipientList.vue
@@ -8,8 +8,8 @@
     :required="required"
     :helper-text="helperText"
     :is-invalid="isInvalid"
-    :no-items-message="$t('TRANSACTION.MULTI_PAYMENT.NO_RECIPIENTS')"
-    class="TransactionMultiPaymentList"
+    :no-items-message="$t('TRANSACTION.NO_RECIPIENTS')"
+    class="TransactionRecipientList"
     @remove="emitRemove"
   >
     <div
@@ -23,7 +23,7 @@
       />
 
       <div class="flex-1 px-4">
-        <div class="TransactionMultiPaymentList__recipient flex py-1">
+        <div class="TransactionRecipientList__recipient flex py-1">
           <span class="font-bold mr-1">
             {{ $t('TRANSACTION.RECIPIENT') }}:
           </span>
@@ -41,12 +41,19 @@
           </span>
         </div>
 
-        <div class="TransactionMultiPaymentList__amount flex py-1">
+        <div class="TransactionRecipientList__amount flex py-1">
           <span class="font-bold mr-1">
             {{ $t('TRANSACTION.AMOUNT') }}:
           </span>
 
-          <span>
+          <span
+            v-if="item.sendAll"
+            class="font-bold uppercase"
+          >
+            {{ $t('COMMON.ALL') }}
+          </span>
+
+          <span v-else>
             {{ formatter_networkCurrency(item.amount) }}
           </span>
         </div>
@@ -62,7 +69,7 @@ import WalletAddress from '@/components/Wallet/WalletAddress'
 import WalletIdenticon from '@/components/Wallet/WalletIdenticon'
 
 export default {
-  name: 'TransactionMultiPaymentList',
+  name: 'TransactionRecipientList',
 
   components: {
     InputEditableList,
@@ -75,7 +82,7 @@ export default {
       type: String,
       required: false,
       default: function () {
-        return this.$t('TRANSACTION.MULTI_PAYMENT.RECIPIENTS')
+        return this.$t('TRANSACTION.RECIPIENTS')
       }
     },
