+++ src/renderer/components/Transaction/TransactionShow/TransactionShow.vue
@@ -232,7 +232,7 @@
         :label="$t('TRANSACTION.RECIPIENTS')"
         item-value-class="items-center"
       >
-        <TransactionMultiPaymentList
+        <TransactionRecipientList
           :title="null"
           :items="transaction.asset.payments"
           :show-links="true"
@@ -263,7 +263,7 @@ import { ListDivided, ListDividedItem } from '@/components/ListDivided'
 import { ModalWindow } from '@/components/Modal'
 import { ButtonClipboard, ButtonGeneric } from '@/components/Button'
 import SvgIcon from '@/components/SvgIcon'
-import { TransactionAmount, TransactionMultiPaymentList, TransactionStatusIcon } from '@/components/Transaction'
+import { TransactionAmount, TransactionRecipientList, TransactionStatusIcon } from '@/components/Transaction'
 import WalletAddress from '@/components/Wallet/WalletAddress'
 import WalletService from '@/services/wallet'
 import truncateMiddle from '@/filters/truncate-middle'
@@ -279,7 +279,7 @@ export default {
     ButtonClipboard,
     SvgIcon,
     TransactionAmount,
-    TransactionMultiPaymentList,
+    TransactionRecipientList,
     TransactionStatusIcon,
     WalletAddress
   },
