+++ src/renderer/components/Transaction/index.js
@@ -1,8 +1,8 @@
 import TransactionAmount from './TransactionAmount'
 import TransactionDetail from './TransactionDetail'
 import TransactionModal from './TransactionModal'
-import TransactionMultiPaymentList from './TransactionMultiPaymentList'
 import TransactionPeerList from './TransactionPeerList'
+import TransactionRecipientList from './TransactionRecipientList'
 import TransactionStatusIcon from './TransactionStatusIcon'
 
 export * from './TransactionForm'
@@ -11,8 +11,8 @@ export {
   TransactionAmount,
   TransactionDetail,
   TransactionModal,
-  TransactionMultiPaymentList,
   TransactionPeerList,
+  TransactionRecipientList,
   TransactionStatusIcon
 }
