+++ src/renderer/i18n/locales/en-US.js
@@ -1029,6 +1029,7 @@ export default {
       LOAD_FROM_FILE: 'Failed to load transaction file',
       EXPIRED: 'Transaction expired before it was processed: {transactionId}',
       FEE_TOO_LOW: 'Transaction could not be sent because the fee ({fee}) is too low',
+      NOT_ENOUGH_BALANCE: 'The total amount of the recipients + fee is greater than the balance ({balance}) of the wallet',
       NOTHING_SENT: 'The transaction could not be sent. Please check your network connection or change peer',
       WRONG_NONCE: 'The transaction could not be sent. Wrong nonce provided',
       NETWORK_NOT_CONFIGURED: 'Network not configured',
@@ -1072,9 +1073,6 @@ export default {
       UNVOTE_DELEGATE: 'Unvote delegate {delegate}'
     },
     MULTI_PAYMENT: {
-      BUTTON_ADD: 'Add',
-      NO_RECIPIENTS: 'There are no recipients',
-      RECIPIENTS: 'Recipients',
       TOTAL_AMOUNT: 'Total amount',
       WARNING_DUPLICATE: 'The address is already a recipient'
     },
@@ -1116,10 +1114,11 @@ export default {
     },
     AMOUNT: 'Amount',
     BLOCK_ID: 'Block ID',
-    BUSINESS_NAME: 'Business Name',
     BRIDGECHAIN_NAME: 'Bridgechain Name',
+    BUSINESS_NAME: 'Business Name',
+    BUTTON_ADD: 'Add',
     CONFIRMATION_COUNT: '{confirmations} Confirmations',
-    CONFIRM_SEND_ALL: 'This will enable sending all of your tokens from the current wallet in this transaction.',
+    CONFIRM_SEND_ALL: 'This will enable sending all of your tokens from the current wallet in this transaction. The selected transfer fee will be deducted from the send amount.',
     CONFIRM_SEND_ALL_TITLE: 'Send all your tokens?',
     CONFIRM_SEND_ALL_NOTE: 'Note: once sent, this cannot be undone.',
     CONFIRMATIONS: 'Confirmations',
@@ -1136,6 +1135,7 @@ export default {
     LOAD_FROM_FILE: 'Load transaction from file',
     MULTIPLE: 'Multiple',
     NONCE: 'Nonce',
+    NO_RECIPIENTS: 'There are no recipient(s)',
     OPEN_IN_EXPLORER: 'Open in Explorer',
     PASSPHRASE: 'Passphrase',
     PASSWORD: 'Encryption Password',
